f08_后端/p114_grpc/requirements.txt
# grpc
grpcio
grpcio-tools